Extract handleErrorPopup method
