+++ src/web/utils/apiError.tsx
@@ -1,4 +1,6 @@
 import { AxiosError } from 'axios';
+
+import { StatusNotificationType } from '../components/Core/StatusPopup';
 /*
  * Backend errors may return an `errorHash` that can help correlate the logs
  * with the user reported issues.
@@ -24,4 +26,19 @@ export function backendError(e: unknown, overrideMessage: string) {
     });
   }
   return Error(overrideMessage);
+}
+
+export function handleErrorPopup(
+  e: Error,
+  setStatusPopup: React.Dispatch<React.SetStateAction<StatusNotificationType | undefined>>,
+  setShowStatusPopup: React.Dispatch<React.SetStateAction<boolean>>
+) {
+  const hasHash = Object.hasOwn(e, 'errorHash') && (e as ApiError).errorHash;
+  const hash = hasHash ? `: (${(e as ApiError).errorHash})` : '';
+  setStatusPopup({
+    type: 'Error',
+    message: `${e.message}${hash}`,
+  });
+  setShowStatusPopup(true);
+  throw new Error(e.message);
 }
